requirements.txt
requests==2.33.1
redis==7.4.0
python-dotenv==1.2.2
PyYAML==6.0.3
boto3==1.42.85
joblib==1.5.3
scikit-learn==1.7.2
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyzmq==27.1.0
SQLAlchemy==2.0.49